requirements.txt
click==8.1.7
requests==2.31.0
pyyaml==6.0.1
